+++ src/tool/cmd.cpp
@@ -347,18 +347,6 @@ std::string cmd::strVal(const std::string &option) const
     }
 }
 
-std::int64_t cmd::int64Val(const std::string &option) const
-{
-    auto opt = this->opt(option);
-    return opt.set ? ::atoll(opt.val.c_str()) : static_cast<std::int64_t>(opt.def);
-}
-
-double cmd::doubleVal(const std::string &option) const
-{
-    auto opt = this->opt(option);
-    return opt.set ? std::atof(opt.val.c_str()) : static_cast<double>(opt.def);
-}
-
 // set
 bool cmd::isSet(const std::string &option) const
 {